+++ src/driver/sqlx_postgres.rs
@@ -36,6 +36,21 @@ impl std::fmt::Debug for SqlxPostgresPoolConnection {
     }
 }
 
+impl From<PgPool> for SqlxPostgresPoolConnection {
+    fn from(pool: PgPool) -> Self {
+        SqlxPostgresPoolConnection {
+            pool,
+            metric_callback: None,
+        }
+    }
+}
+
+impl From<PgPool> for DatabaseConnection {
+    fn from(pool: PgPool) -> Self {
+        DatabaseConnection::SqlxPostgresPoolConnection(pool.into())
+    }
+}
+
 impl SqlxPostgresConnector {
     /// Check if the URI provided corresponds to `postgres://` for a PostgreSQL database
     pub fn accepts(string: &str) -> bool {